add core utils, combine utility methods
