+++ src/index.ts
@@ -7,5 +7,4 @@ export * from './utils/lf-localization.service.js';
 export * as MimeTypeUtils from './utils/mime-type-utils.js';
 export * as PathUtils from './utils/path-utils.js';
 export * as StringUtils from './utils/string-utils.js';
-export * as TimeUtils from './utils/time-utils.js';
-export * from './utils/validate-defined.js';
+export * as CoreUtils from './utils/core-utils.js';
